requirements.txt
Pillow==9.5.0
Pillow==9.5.0
scikit_learn==1.2.2
streamlit==1.24.0
